Add files via upload
This website was made for a client who is in the "Salon Business", it its like the client's portfolio where the Client's users will be able to visit this website to see the client's work. it also gives the client's user the option to chat to the client on making a suggestion on what hairstle the user would like to do but before the user gets to chat with the our client, the client will have to login.